Move the bootloader-related code to the module
Remove the Python package pyanaconda.bootloader.
