+++ pyanaconda/modules/storage/bootloader/image.py
