+++ pyanaconda/modules/storage/bootloader/zipl.py
@@ -18,7 +18,7 @@
 import os
 import re
 
-from pyanaconda.bootloader.base import BootLoader, Arguments, BootLoaderError
+from pyanaconda.modules.storage.bootloader.base import BootLoader, Arguments, BootLoaderError
 from pyanaconda.core import util
 from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.product import productName
