+++ pyanaconda/modules/storage/partitioning/automatic/noninteractive_partitioning.py
@@ -21,7 +21,7 @@ from blivet.errors import NoDisksError
 from blivet.formats.disklabel import DiskLabel
 from pyanaconda.anaconda_loggers import get_module_logger
 
-from pyanaconda.bootloader.execution import setup_bootloader
+from pyanaconda.modules.storage.bootloader.execution import setup_bootloader
 from pyanaconda.modules.common.constants.objects import DISK_INITIALIZATION
 from pyanaconda.modules.common.constants.services import STORAGE
 from pyanaconda.modules.storage.disk_initialization import DiskInitializationConfig
