+++ pyanaconda/storage/osinstall.py
@@ -30,7 +30,7 @@ from blivet.size import Size
 from blivet.devicelibs.crypto import DEFAULT_LUKS_VERSION
 
 from pyanaconda.core import util
-from pyanaconda.bootloader import BootLoaderFactory
+from pyanaconda.modules.storage.bootloader import BootLoaderFactory
 from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.core.constants import shortProductName
 from pyanaconda.storage.fsset import FSSet
